# Optimize scan method