@naitoh 2024-10-15
@kou 
> Could you add how to optimize this to the description? 

Sorry.
I added it to the description.