@kou 2024-10-15
Thanks.
How about using 1 PR for 1 optimization?

It seems that this has 4 optimizations:

1. CRuby: `(S_RESTLEN(p) < RSTRING_LEN(pattern))`
2. CRuby: `rb_enc_get()`
3. JRuby: `(restLen() < pattern.size())`
4. JRuby: `currPtr()`